Fix HTML code bug in proofreader
Literal HTML code in forum posts would be removed from the preview by
the proofreader. For example, the post

    You can use the <span> tag

would become

    You can use the tag

when previewing.

The processed string must be inserted as innerHTML, because the
highlighting itself is done using HTML, so the only solution I could
come up with was to escape the input string first.